+++ spec/spec_helper.rb
@@ -3,5 +3,5 @@ require 'chefspec/berkshelf'
 
 RSpec.configure do |config|
   config.platform = 'ubuntu'
-  config.version = '14.04'
+  config.version = '16.04'
 end