+++ test/cookbooks/test/metadata.rb
@@ -2,4 +2,4 @@ name 'test'
 version '0.0.1'
 
 depends 'stunnel'
-depends 'nginx', '~> 10.0.2'
+depends 'nginx'