+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ## 4.1.3 - *2023-02-14*
 
 - Remove delivery folder
+- Fix test cookbook by updating nginx dependency & creating an nginx_service resource
+- Add unified_mode true & String type to connect & accept properties for connection resource
+- Split certs into server & client in test cookbook
 
 ## 4.1.2 - *2021-08-31*